Rename io.py to stio.py due to import error
When nlp.py is executed, it makes following error.
'module' object has no attribute 'BufferedIOBase'

It means that it calls io.py instead of system io package so that
that kind of BufferedIOBase problem. Of course, renaming io.py is
not good solution to resolve this problem, but temporary solution.

FIXME: fix this problem.